Feature: Clean up code base (#40)
* fix auto fixable ESLint warnings

* use .jsx file extension

* move Header into it's own component

* use .js insteda of .jsx file extension

* format code using prettier
