+++ gatsby-config.js
@@ -1,6 +1,6 @@
 module.exports = {
   siteMetadata: {
-    title: `Gatsby Default Starter`,
+    title: 'Gatsby Default Starter',
   },
-  plugins: [`gatsby-plugin-react-helmet`],
-}
+  plugins: ['gatsby-plugin-react-helmet'],
+};
